# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "api/articals": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "list articles",
                "operationId": "App\\Http\\Controllers\\ArticalController::index",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "A list with articals"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "add new article",
                "operationId": "addArticle",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Article added successfuly"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "api/articals/{articleID}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Find aericle by ID",
                "operationId": "App\\Http\\Controllers\\ArticalController::show",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "ID of article that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Article"
                ],
                "summary": "Update data for rticle",
                "operationId": "updateArticle",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "ID of article that needs to be update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Article added successsfuly"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "Delete an article",
                "operationId": "DeleteArticle",
                "parameters": [
                    {
                        "name": "authorId",
                        "in": "path",
                        "description": "ID of article that needs to be update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid "
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            }
        },
        "/login": {
            "get": {
                "tags": [
                    "Author Login"
                ],
                "summary": "Logs author into the system",
                "description": "",
                "operationId": "loginAuthor",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The user email for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The user password for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    }
                }
            }
        },
        "api/authors": {
            "get": {
                "tags": [
                    "Author"
                ],
                "summary": "list authors",
                "operationId": "App\\Http\\Controllers\\AuthorController::index",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "A list with authors"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "api/authors/{authorID}": {
            "get": {
                "tags": [
                    "Author"
                ],
                "summary": "Find author by ID",
                "operationId": "App\\Http\\Controllers\\AuthorController::show",
                "parameters": [
                    {
                        "name": "authorId",
                        "in": "path",
                        "description": "ID of author that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Author not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Author"
                ],
                "summary": "Update data for author",
                "operationId": "updateAuthor",
                "parameters": [
                    {
                        "name": "authorId",
                        "in": "path",
                        "description": "ID of author that needs to be update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Author added successsfuly"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Author"
                ],
                "summary": "Delete an author",
                "operationId": "DeleteAuthor",
                "parameters": [
                    {
                        "name": "authorId",
                        "in": "path",
                        "description": "ID of author that needs to be update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid "
                    },
                    "404": {
                        "description": "Author not found"
                    }
                }
            }
        },
        "/authors": {
            "post": {
                "tags": [
                    "Author"
                ],
                "summary": "add new author",
                "operationId": "addAuthor",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Author added successfuly"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Artical": {
                "properties": {
                    "main_title": {
                        "type": "string"
                    },
                    "secondary_title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "author_id": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Author"
                }
            }
        }
    }
}